+++ src/icons/icon.component.jsx
@@ -2,158 +2,231 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import IndicatorRemove from
-  '../../images/indicators/Remove.svg';
-import IndicatorBurgerClose from
-  '../../images/indicators/Burger-close.svg';
-import IndicatorBurger from
-  '../../images/indicators/Burger.svg';
-import IndicatorLogout from
-  '../../images/indicators/Logout.svg';
-import IndicatorCaretRight from
-  '../../images/indicators/CaretRight.svg';
-import IndicatorCaretLeft from
-  '../../images/indicators/CaretLeft.svg';
-import IndicatorFlagged from
-  '../../images/indicators/Flagged.svg';
-import IndicatorComment from
-  '../../images/indicators/Comment.svg';
-import IndicatorCommented from
-  '../../images/indicators/Commented.svg';
-import IndicatorEmail from
-  '../../images/indicators/Email.svg';
-import IndicatorAttachment from
-  '../../images/indicators/Attachment.svg';
-import IndicatorInClarification from
-  '../../images/indicators/StatusInclarification.svg';
-import IndicatorRejected from
-  '../../images/indicators/Rejected.svg';
-import IndicatorLocked from
-  '../../images/indicators/StatusLocked.svg';
-import IndicatorOk from
-  '../../images/indicators/Ok.svg';
-import IndicatorInspected from
-  '../../images/indicators/Inspected.svg';
-import IndicatorInspectedAndApproved from
-  '../../images/indicators/InspectedAndApproved.svg';
-import IndicatorClose from
-  '../../images/indicators/Close.svg';
-import IndicatorDelete from
-  '../../images/indicators/Delete.svg';
 import IndicatorAlert from
-  '../../images/indicators/Alert.svg';
-import IndicatorWarning from
-  '../../images/indicators/Warning.svg';
-import IndicatorError from
-  '../../images/indicators/Error.svg';
-import IndicatorCritical from
-  '../../images/indicators/Critical.svg';
-import IndicatorExclamation from
-  '../../images/indicators/Exclamation.svg';
-import IndicatorPlus from
-  '../../images/indicators/Plus.svg';
-import IndicatorMinus from
-  '../../images/indicators/Minus.svg';
+  './images/indicators/Alert.svg';
 import IndicatorArrowLeft from
-  '../../images/indicators/ArrowLeft.svg';
-import IndicatorHelp from
-  '../../images/indicators/Help.svg';
-import IndicatorMore from
-  '../../images/indicators/More.svg';
-import IndicatorSettings from
-  '../../images/indicators/Settings.svg';
-import IndicatorSearch from
-  '../../images/indicators/Search.svg';
-import IndicatorSearchWithCircle from
-  '../../images/indicators/SearchWithCircle.svg';
-import IndicatorPinned from
-  '../../images/indicators/Pinned.svg';
+  './images/indicators/ArrowLeft.svg';
+import IndicatorAttachment from
+  './images/indicators/Attachment.svg';
+
+import IndicatorBurger from
+  './images/indicators/Burger.svg';
+import IndicatorBurgerClose from
+  './images/indicators/Burger-close.svg';
+
+import IndicatorCalendar from
+  './images/indicators/Calendar.svg';
+import IndicatorCaretLeft from
+  './images/indicators/CaretLeft.svg';
+import IndicatorCaretRight from
+  './images/indicators/CaretRight.svg';
+import IndicatorClock from
+  './images/indicators/Clock.svg';
+import IndicatorClose from
+  './images/indicators/Close.svg';
+import IndicatorComment from
+  './images/indicators/Comment.svg';
+import IndicatorCommented from
+  './images/indicators/Commented.svg';
+import IndicatorCritical from
+  './images/indicators/Critical.svg';
+
+import IndicatorDelete from
+  './images/indicators/Delete.svg';
 import IndicatorDraggingArrows from
-  '../../images/indicators/DraggingArrows.svg';
+  './images/indicators/DraggingArrows.svg';
+
+import IndicatorEdit from
+  './images/indicators/Edit.svg';
+import IndicatorEmail from
+  './images/indicators/Email.svg';
+import IndicatorError from
+  './images/indicators/Error.svg';
+import IndicatorExclamation from
+  './images/indicators/Exclamation.svg';
+import IndicatorExport from
+  './images/indicators/Export.svg';
+
+import IndicatorFiltering from
+  './images/indicators/Filtering.svg';
+import IndicatorFlagged from
+  './images/indicators/Flagged.svg';
+import IndicatorFolder from
+  './images/indicators/Folder.svg';
+
+import IndicatorHelp from
+  './images/indicators/Help.svg';
+
+import IndicatorInspected from
+  './images/indicators/Inspected.svg';
+import IndicatorInspectedAndApproved from
+  './images/indicators/InspectedAndApproved.svg';
+
+import IndicatorLogout from
+  './images/indicators/Logout.svg';
+
+import IndicatorMinus from
+  './images/indicators/Minus.svg';
+import IndicatorMore from
+  './images/indicators/More.svg';
+
+import IndicatorOk from
+  './images/indicators/Ok.svg';
+
+import IndicatorPinned from
+  './images/indicators/Pinned.svg';
+import IndicatorPlus from
+  './images/indicators/Plus.svg';
+import IndicatorPresentation from
+  './images/indicators/Presentation.svg';
+import IndicatorPrint from
+  './images/indicators/Print.svg';
+
+import IndicatorRejected from
+  './images/indicators/Rejected.svg';
+import IndicatorRemove from
+  './images/indicators/Remove.svg';
+import RemoveFilter from
+  './images/indicators/RemoveFilter.svg';
+
+import IndicatorSave from
+  './images/indicators/Save.svg';
+import IndicatorSearch from
+  './images/indicators/Search.svg';
+import IndicatorSearchWithCircle from
+  './images/indicators/SearchWithCircle.svg';
+import IndicatorSettings from
+  './images/indicators/Settings.svg';
+import IndicatorSortAsc from
+  './images/indicators/SortAsc.svg';
+import IndicatorSortDesc from
+  './images/indicators/SortDesc.svg';
+import IndicatorInClarification from
+  './images/indicators/StatusInclarification.svg';
+import IndicatorLocked from
+  './images/indicators/StatusLocked.svg';
+import IndicatorStatusUser from
+  './images/indicators/StatusUser.svg';
+
+import IndicatorWarning from
+  './images/indicators/Warning.svg';
+import IndicatorExceedsLimit from
+  './images/indicators/ExceedsLimit.svg';
 
 import ProductInvoices from
-  '../../images/products/Invoices.svg';
+  './images/products/Invoices.svg';
 import ProductUser from
-  '../../images/products/User.svg';
+  './images/products/User.svg';
 import ProductAccounts from
-  '../../images/products/Accounts.svg';
+  './images/products/Accounts.svg';
 import ProductArchive from
-  '../../images/products/Archive.svg';
+  './images/products/Archive.svg';
 import ProductDashboard from
-  '../../images/products/Dashboard.svg';
+  './images/products/Dashboard.svg';
 import ProductInfo from
-  '../../images/products/Info.svg';
+  './images/products/Info.svg';
 import ProductLiquidity from
-  '../../images/products/Liquidity.svg';
+  './images/products/Liquidity.svg';
 import ProductNetting from
-  '../../images/products/Netting.svg';
+  './images/products/Netting.svg';
 import ProductPayments from
-  '../../images/products/Payments.svg';
+  './images/products/Payments.svg';
 import ProductSettings from
-  '../../images/products/Settings.svg';
+  './images/products/Settings.svg';
 import ProductTools from
-  '../../images/products/Tools.svg';
+  './images/products/Tools.svg';
 import ProductUsers from
-  '../../images/products/Users.svg';
+  './images/products/Users.svg';
 import ProductMatching from
-  '../../images/products/Matching.svg';
+  './images/products/Matching.svg';
 import ProductInspector from
-  '../../images/products/Inspector.svg';
+  './images/products/Inspector.svg';
 
 import InvoicesGeneric from
-  '../../images/invoices/document/generic.svg';
+  './images/invoices/document/generic.svg';
 import InvoicesPdf from
-  '../../images/invoices/document/pdf.svg';
+  './images/invoices/document/pdf.svg';
 import InvoicesPng from
-  '../../images/invoices/document/png.svg';
+  './images/invoices/document/png.svg';
 import InvoicesPpt from
-  '../../images/invoices/document/ppt.svg';
+  './images/invoices/document/ppt.svg';
 import InvoicesWord from
-  '../../images/invoices/document/word.svg';
+  './images/invoices/document/word.svg';
 import InvoicesXls from
-  '../../images/invoices/document/xls.svg';
+  './images/invoices/document/xls.svg';
 
 import OCLong from
-  '../../images/logo/oc-logo.svg';
+  './images/logo/oc-logo.svg';
 import OCShort from
-  '../../images/logo/oc-logo-short.svg';
+  './images/logo/oc-logo-short.svg';
 
 const components = {
   indicator: {
-    burgerClose: React.createFactory(IndicatorBurgerClose),
-    remove: React.createFactory(IndicatorRemove),
+    alert: React.createFactory(IndicatorAlert),
+    arrowLeft: React.createFactory(IndicatorArrowLeft),
+    attachment: React.createFactory(IndicatorAttachment),
+
     burger: React.createFactory(IndicatorBurger),
-    logout: React.createFactory(IndicatorLogout),
+    burgerClose: React.createFactory(IndicatorBurgerClose),
+
+    calendar: React.createFactory(IndicatorCalendar),
     CaretRight: React.createFactory(IndicatorCaretRight),
     CaretLeft: React.createFactory(IndicatorCaretLeft),
-    flagged: React.createFactory(IndicatorFlagged),
+    clock: React.createFactory(IndicatorClock),
+    close: React.createFactory(IndicatorClose),
     comment: React.createFactory(IndicatorComment),
     commented: React.createFactory(IndicatorCommented),
+    critical: React.createFactory(IndicatorCritical),
+
+    delete: React.createFactory(IndicatorDelete),
+    draggingArrows: React.createFactory(IndicatorDraggingArrows),
+
+    edit: React.createFactory(IndicatorEdit),
     email: React.createFactory(IndicatorEmail),
-    attachment: React.createFactory(IndicatorAttachment),
-    inClarification: React.createFactory(IndicatorInClarification),
-    rejected: React.createFactory(IndicatorRejected),
-    locked: React.createFactory(IndicatorLocked),
-    ok: React.createFactory(IndicatorOk),
+    error: React.createFactory(IndicatorError),
+    exclamation: React.createFactory(IndicatorExclamation),
+    export: React.createFactory(IndicatorExport),
+
+    filtering: React.createFactory(IndicatorFiltering),
+    flagged: React.createFactory(IndicatorFlagged),
+    folder: React.createFactory(IndicatorFolder),
+
+    help: React.createFactory(IndicatorHelp),
+
     inspected: React.createFactory(IndicatorInspected),
     inspectedAndApproved: React.createFactory(IndicatorInspectedAndApproved),
-    close: React.createFactory(IndicatorClose),
-    delete: React.createFactory(IndicatorDelete),
-    alert: React.createFactory(IndicatorAlert),
-    warning: React.createFactory(IndicatorWarning),
-    error: React.createFactory(IndicatorError),
-    critical: React.createFactory(IndicatorCritical),
-    exclamation: React.createFactory(IndicatorExclamation),
-    plus: React.createFactory(IndicatorPlus),
+
+    logout: React.createFactory(IndicatorLogout),
+
     minus: React.createFactory(IndicatorMinus),
-    arrowLeft: React.createFactory(IndicatorArrowLeft),
-    help: React.createFactory(IndicatorHelp),
     more: React.createFactory(IndicatorMore),
-    settings: React.createFactory(IndicatorSettings),
+
+    ok: React.createFactory(IndicatorOk),
+
+    pinned: React.createFactory(IndicatorPinned),
+    plus: React.createFactory(IndicatorPlus),
+    presentation: React.createFactory(IndicatorPresentation),
+    print: React.createFactory(IndicatorPrint),
+
+    rejected: React.createFactory(IndicatorRejected),
+    remove: React.createFactory(IndicatorRemove),
+    removeFilter: React.createFactory(RemoveFilter),
+
+    save: React.createFactory(IndicatorSave),
     search: React.createFactory(IndicatorSearch),
     searchWithCircle: React.createFactory(IndicatorSearchWithCircle),
-    pinned: React.createFactory(IndicatorPinned),
-    draggingArrows: React.createFactory(IndicatorDraggingArrows),
+    settings: React.createFactory(IndicatorSettings),
+    sortAsc: React.createFactory(IndicatorSortAsc),
+    sortDesc: React.createFactory(IndicatorSortDesc),
+    inClarification: React.createFactory(IndicatorInClarification),
+    locked: React.createFactory(IndicatorLocked),
+
+    user: React.createFactory(IndicatorStatusUser),
+
+    warning: React.createFactory(IndicatorWarning),
+
+    exceedsLimit: React.createFactory(IndicatorExceedsLimit),
+
   },
   product: {
     Invoices: React.createFactory(ProductInvoices),
@@ -190,14 +263,26 @@ export default class Icon extends React.Component {
   constructor(props) {
     super(props);
     switch (props.type) {
-      case 'product':
+      case 'product': {
         this.defaultWidth = 40;
         this.defaultHeight = 40;
         break;
-      case 'indicator':
-        this.defaultWidth = 30;
-        this.defaultHeight = 30;
+      }
+      case 'indicator': {
+        this.defaultWidth = 25;
+        this.defaultHeight = 25;
         break;
+      }
+      case 'logo': {
+        if (props.name === 'OCLong') {
+          this.defaultWidth = 150;
+          this.defaultHeight = 30;
+        } else {
+          this.defaultWidth = 50;
+          this.defaultHeight = 30;
+        }
+        break;
+      }
       default:
         this.defaultWidth = 40;
         this.defaultHeight = 40;
@@ -227,8 +312,8 @@ export default class Icon extends React.Component {
 }
 
 Icon.defaultProps = {
-  width: 40,
-  height: 40,
+  width: null,
+  height: null,
 };
 
 Icon.propTypes = {