+++ stories/index.story.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { text, number, select } from '@storybook/addon-knobs';
+import { number, select } from '@storybook/addon-knobs';
 import { Icon } from '../src/index';
 
 import './react-icons-story.scss';