+++ test/icons/icon.component.spec.jsx
@@ -11,14 +11,12 @@ describe('Icon component', function describe() {
   it('should render correctly', () => {
     // This will only test the test stub defined in test.setup.js
     // Testing the real Icon component hasn't succeeded yet.
-    const wrapper = mount(
-      <Icon type="indicator" name="more" />,
-    );
+    const wrapper = mount(<Icon type="indicator" name="more" />);
     expect(wrapper.props()).to.eql({
-      height: 40,
+      height: null,
       type: 'indicator',
       name: 'more',
-      width: 40,
+      width: null,
     });
   });
 });