+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
+# 4.0.4
+* Add proper default sizes to logos
+* Improve demo
+
+# 4.0.3
+* Fix svg images with ES build
+
+# 4.0.2
+* Fix svg image to support react-svg-loader 1.1.1
+
+# 4.0.1
+* Fix svg images with ES build
+
+# 4.0.0
+* Update to new component layout
+* Upgrade packages
+
+# 3.0.5
+* Add approval limit exceeds limit indicator
+
+# 3.0.4
+* Upgrade packages
+* Fix tests and linter errors
+
+# 3.0.3
+* Add oc-icon-indicator class to help, inClarification, and Pinned icons
+* Some icons in images/indicators were not included
+
+# 3.0.2
+* Remove title elements from some of the indicator icons
+
+# 3.0.1
+* Add 'remove filter' icon
+
+## 3.0.0
+* Remove margins from indicator icons
+
 ## 2.0.3
 * Update latest version to npm
 
@@ -9,7 +46,7 @@
 * Update size icons: `warning`, `error`,`critical` and `search`
 
 ## 2.0.1
-* Change focusable SVG prop from boolean to string type, fixes React 16 warning 
+* Change focusable SVG prop from boolean to string type, fixes React 16 warning
 
 ## 2.0.0
 * Upgrade to react 15.6 & babel-preset-env